Move LinePool things to sub dir
