+++ Questie-Classic.toc
@@ -222,8 +222,8 @@ Modules\Tracker\TrackerHeaderFrame.lua
 Modules\Tracker\TrackerQuestFrame.lua
 Modules\Tracker\TrackerQuestTimers.lua
 Modules\Tracker\TrackerItemButton.lua
-Modules\Tracker\TrackerLine.lua
-Modules\Tracker\TrackerLinePool.lua
+Modules\Tracker\LinePool\TrackerLine.lua
+Modules\Tracker\LinePool\TrackerLinePool.lua
 
 # Tutorial
 Modules\Tutorial\ShowRunes.lua
